fix(build): prevent prerender crash on /dev/ui
